adding user details route
