+++ api/package.json
@@ -17,6 +17,7 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@charliewilco/sanitize-object": "^1.0.1",
     "@types/bcrypt": "^2.0.0",
     "@types/boom": "^7.2.0",
     "@types/hapi": "^17.0.19",
@@ -72,7 +73,7 @@
     "watch": [
       "./src/**/*.ts"
     ],
-    "exec": "ts-node ./src/server.ts"
+    "exec": "ts-node --no-cache ./src/server.ts"
   },
   "repository": "charlespeters/downwrite",
   "author": "Charles ⚡ <charlespeters42@gmail.com>"
